+++ packages/gatsby/src/bootstrap/index.js
@@ -28,7 +28,7 @@ process.on(`unhandledRejection`, (reason, p) => {
   report.panic(reason)
 })
 
-import { createGraphQLRunner } from "./create-graphql-runner"
+import { createGraphQLRunner } from "./graphql-runner"
 const { extractQueries } = require(`../query/query-watcher`)
 const requiresWriter = require(`./requires-writer`)
 const { writeRedirects } = require(`./redirects-writer`)
